requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
greenlet==1.1.1
idna==3.2
psycopg2-binary==2.9.1
requests==2.26.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
urllib3==1.26.6
python-dotenv==0.19.0